pyproject.toml
[project]
name = "firstclassproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.19",
    "python-decouple>=3.8",
]
